formula/sample_bot_by_model_based/behavior_cloning/requirements.txt
numpy==1.14.5
opencv-python==3.4.2.17
tensorflow==1.10.1
keras==2.2.2
pandas==0.23.4
scipy==1.1.0
h5py==2.8.0
socketio==0.1.3
flask-socketio==1.0b1
eventlet==0.24.1
pillow==5.2.0
Flask==1.0.2
pygame==1.9.4
